% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_age.R
\name{check_age}
\alias{check_age}
\title{Check and Correct Age in Longitudinal Data}
\usage{
check_age(df, Subj_ID, testdate, DOB, Age)
}
\arguments{
\item{df}{the dataframe you would like applied to this function}

\item{Subj_ID}{the column name where patients identifying numbers are stored}

\item{testdate}{the column name where the date the sample was collected is stored}

\item{DOB}{the column name where the patient's date of birth is stored}

\item{Age}{the column where the patient's age is stored}
}
\value{
The entered dataframe now with any incorrectly entered ages fixed
}
\description{
This function checks to ensure a subject's age is consistent over longitudinal data. It will also correct any incorrect ages.
}
\examples{
#This example dataset has an incorrectly entered in the Age column. The third
#TestDate for ID 1 was entered as 8.83, but based on the DOB this patient should only have been 5.83 years old at this visit
example <- data.frame(
ID = c(1, 1, 1, 2, 2),
TestDate = as.Date(c("2023-01-01", "2023-01-01", "2023-04-01", "2023-02-01", "2023-02-01")),
DOB = as.Date(c("2017-06-01", "2017-06-01", "2017-06-01", "2017-12-14", "2017-12-14")),
Age = c(5.58, 5.58, 8.83, 5.05, 5.05))
check_age(example, "ID", "TestDate", "DOB", "Age")
#Notice now, this incorrect age has been corrected
example
}
