% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_visit_order.R
\name{check_and_fix_visit_num_order}
\alias{check_and_fix_visit_num_order}
\title{My Hello World Function}
\usage{
check_and_fix_visit_num_order(df)
}
\arguments{
\item{df}{}
}
\description{
My Hello World Function
}
