% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_visit_order.R
\name{check_visit_order}
\alias{check_visit_order}
\title{Check and Correct Visit Order Number in Longitudinal Data}
\usage{
check_visit_order(df, Subj_ID, testdate, visit_num)
}
\arguments{
\item{df}{the dataframe you would like applied to this function}

\item{Subj_ID}{the patients identifying number}

\item{testdate}{the date the sample was collected}

\item{visit_num}{the visit number, the record of which instance the patient was seen}
}
\value{
The entered dataframe now with any incorrectly entered visit order number information fixed
}
\description{
This function checks to ensure a subject's visit order number is correct over longitudinal data. It will also correct any incorrect entries.
}
\examples{
This example dataset has an incorrectly entered in the visit_num column. The second and third
#visit_num for ID 1 are accidentally reversed.
example <- data.frame(
ID = c(1, 1, 1, 2, 2),
TestDate = as.Date(c("2023-01-01", "2023-02-01", "2023-04-01", "2023-02-01", "2023-02-01")),
visit_num = c(1, 3, 2, 1, 2))
check_visit_order(example, "ID", "TestDate", "DOB", "Age")
#Notice now, the incorrect visit order numbers have been corrected
example
}
