% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_age.R
\docType{data}
\name{Example_Dataset}
\alias{Example_Dataset}
\title{Check Age in longitudinal data}
\format{
An object of class \code{data.frame} with 230 rows and 14 columns.
}
\usage{
Example_Dataset
}
\arguments{
\item{Subj_ID}{the patients identifying number}

\item{df}{the dataframe you would like applied to this function}

\item{testdate}{the date the sample was collected}

\item{DOB}{the patients date of birth}

\item{Age}{the patients age}
}
\value{
The entered dataframe now with any incorrectly entered ages fixed
}
\description{
This function checks to ensure a subject's age is consistant over longitudinal data.
}
\examples{
#This example dataset has an incorrectly entered in the Age column. The third
#TestDate for ID 1 was entered as 8.83, but based on the DOB this patient should only have been 5.83 years old at this visit
example <- data.frame(
ID = c(1, 1, 1, 2, 2),
TestDate = as.Date(c("2023-01-01", "2023-01-01", "2023-04-01", "2023-02-01", "2023-02-01")),
DOB = as.Date(c("2017-06-01", "2017-06-01", "2017-06-01", "2017-12-14", "2017-12-14")),
Age = c(5.58, 5.58, 8.83, 5.05, 5.05)
check_Age(example, "ID", "TestDate", "DOB", "Age")
#Notice now, this incorrect age has been corrected
example
}
\keyword{datasets}
