% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_stage.R
\name{update_stage}
\alias{update_stage}
\title{Update Venous Confirmatory Sample Stage}
\usage{
update_stage(df, Subj_ID, visit_num, Case_Status, Test_Type)
}
\arguments{
\item{df}{the dataframe you would like applied to this function}

\item{Subj_ID}{the column name where patients identifying numbers are stored}

\item{visit_num}{the column the visit number, the record of which instance the patient was seen, is stored}

\item{Case_Status}{the column where the category of your case investigation, typically inital or confirmatory, is stored}

\item{Test_Type}{the column where the source of your sample, for example the collection method used, is stored}
}
\value{
The entered dataframe now with any incorrectly entered case statuses information fixed based on testing results
}
\description{
This function is specific to case management of elevated blood lead patients and how samples are automatically coded as confirmatory if they are venous blood draws even if they are an initial test result.
}
\examples{
This example dataset has an incorrectly entered case status of initial despite a venous blood draw, which automatically confirms the case.
example <- data.frame(
ID = c(1, 1, 1, 2, 2),
visit_num = c(1, 3, 2, 1, 2),
Case_Status = c( "Initial", "Confirmatory", "Most_Recent", "Initial", "Most_Recent"),
Test_Type = c( "Capillary", "Capillary", "Venous" , "Venous", "Capillary"))
update_stage(example, "ID", "visit_num", "Case_Status", "Test_Type")
#Notice now, the incorrect case status has been corrected
example
}
