% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_stage.R
\name{update_venous_specimen_stage}
\alias{update_venous_specimen_stage}
\title{My Hello World Function}
\usage{
update_venous_specimen_stage(df)
}
\arguments{
\item{df}{}
}
\description{
My Hello World Function
}
